+++ app/models/project.rb
@@ -91,7 +91,7 @@ class Project < ApplicationRecord
 
   # Returns true if the languages limit has been reached.
   def max_languages_reached?
-    organization_plan = self.organization.plan
+    organization_plan = self.organization.current_plan
     organization_plan&.languages_limit.nil? ? false : self.languages.size >= organization_plan.languages_limit
   end