+++ db/migrate/20230121000241_add_keys_count_to_organizations.rb
@@ -3,8 +3,10 @@ class AddKeysCountToOrganizations < ActiveRecord::Migration[6.1]
     add_column :organizations, :keys_count, :integer, default: 0, null: false
 
     Organization.all.each do |organization|
-      organization.keys_count = organization.projects.reduce(0) { |sum, project| sum + project.keys.size }
-      organization.save!
+      organization.update_column(
+        :keys_count,
+        organization.projects.reduce(0) { |sum, project| sum + project.keys.size }
+      )
     end
   end
 end